oh...one more thing*.
auto-download is in. now if you have both auto-check and auto-download enabled you pretty much only have to run CNU+ for the goodness to work.

there is 1 known bug relating to enabling/disabled the auto-download feature then re-checking for updates in the same session. the temporary solution is to just restart CNU+ after changing that setting.

*rest in peace, steve jobs